requirements.txt
flask
python-dotenv
requests
weasyprint
flask_weasyprint